new changes

+++ products/models.py
@@ -8,15 +8,13 @@ class Category(models.Model):
         def __str__(self):
                 return self.name
         
-
-
 class Product(models.Model):
-        category=models.ForeignKey(Category,on_delete=models.CASCADE, related_name='products')
+        category=models.ForeignKey(Category, on_delete=models.CASCADE, related_name='products')
         name=models.CharField(max_length=100)
         description=models.TextField()
         price=models.DecimalField(max_digits=10,decimal_places=2)
-        stock=models.PositiveIntegerField()
+        stock=models.PositiveBigIntegerField()
         image=models.ImageField(upload_to='products/')
 
         def __str__(self):
-                  return self.name
+                return self.name
